feat: Split passive trees into each class
