+++ Content/Items/Gear/Enums.cs
@@ -58,3 +58,13 @@
 	 Rare = 2,
 	 Unique = 3
  }
+ 
+ [Flags]
+ public enum PlayerClass
+ {
+	 None = 0,
+	 Melee = 1,
+	 Ranged = 2,
+	 Magic = 3,
+	 Summoner = 4
+ }
