+++ Content/Passives/RangedPassives.cs
@@ -1,14 +1,16 @@
-﻿using PathOfTerraria.Core.Systems.TreeSystem;
+﻿using PathOfTerraria.Content.Items.Gear;
+using PathOfTerraria.Core.Systems.TreeSystem;
 
 namespace PathOfTerraria.Content.Passives;
 
 internal class AmmoPassive : Passive
 {
-	public AmmoPassive() : base()
+	public AmmoPassive()
 	{
 		Name = "Secret Compartment";
 		Tooltip = "5% chance to not consume ammo per level";
 		MaxLevel = 3;
 		TreePos = new Vector2(300, 180);
+		Classes = [PlayerClass.Ranged];
 	}
 }
