+++ Content/Passives/StartPassive.cs
@@ -1,11 +1,12 @@
 ﻿using PathOfTerraria.Core.Systems.TreeSystem;
 using System.Collections.Generic;
+using PathOfTerraria.Content.Items.Gear;
 
 namespace PathOfTerraria.Content.Passives;
 
 internal class StartPassive : Passive
 {
-	public StartPassive() : base()
+	public StartPassive()
 	{
 		Name = "Anchor";
 		Tooltip = "Your journey starts here";
@@ -15,6 +16,7 @@ internal class StartPassive : Passive
 		Width = 58;
 		Height = 58;
 		TreePos = new Vector2(400, 500);
+		Classes = [PlayerClass.Melee, PlayerClass.Magic, PlayerClass.Ranged, PlayerClass.Summoner];
 	}
 
 	public override void Connect(List<Passive> all, Player player)
